+++ app/Http/Dataproviders/Interfaces/FilterableDataproviderInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Http\Dataproviders\Interfaces;
+
+interface FilterableDataproviderInterface
+{
+    public function getFilterList(): array;
+}
