+++ routes/web.php
@@ -1,17 +1,16 @@
 <?php
 
-use App\Dataproviders\Cardlists\Config\ProjectOverviewCardlist;
-use App\Dataproviders\Cardlists\Modules\PKSanc\PKSancOverviewCardList;
-use App\Dataproviders\Cardlists\Modules\PKSanc\PKSancPokedexCardList;
-use App\Dataproviders\Cardlists\Modules\PKSanc\PKSancStagingCardList;
-use App\Dataproviders\SelectorLists\Modules\PKSanc\FilterSelects\OwnedPokemonSpecies;
-use App\Dataproviders\SelectorLists\Modules\PKSanc\GameDataSelect;
 use App\Http\Controllers\Auth\AuthController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\Modules\PKSanc\PKSancContributionController;
 use App\Http\Controllers\Modules\PKSanc\PKSancController;
 use App\Http\Controllers\Modules\PKSanc\PKSancDepositController;
-use App\Http\Controllers\Modules\PKSanc\PKSancPokdexController;
+use App\Http\Dataproviders\Cardlists\Config\ProjectOverviewCardlist;
+use App\Http\Dataproviders\Cardlists\Modules\PKSanc\PKSancOverviewCardList;
+use App\Http\Dataproviders\Cardlists\Modules\PKSanc\PKSancPokedexCardList;
+use App\Http\Dataproviders\Cardlists\Modules\PKSanc\PKSancStagingCardList;
+use App\Http\Dataproviders\SelectorLists\Modules\PKSanc\FilterSelects\OwnedPokemonSpecies;
+use App\Http\Dataproviders\SelectorLists\Modules\PKSanc\GameDataSelect;
 use Illuminate\Support\Facades\Route;
 
 //| home
